+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 0.3.3.0
+
+* Remove `Reframer`
+* Add `SyncStream` type
+* Add functions to add a sequence number and timestamp to a `SyncStream`
+  rendering it a (reqular) `Stream`
+
 # 0.3.2.0
 
 * Add utility module `Reframer` for _timestamp_ adaption of 'Frame's